+++ src/image/sub_image.rs
@@ -32,6 +32,10 @@ where
 
         Self { parent, area }
     }
+
+    pub(crate) fn new_unchecked(parent: &'a T, area: Rectangle) -> Self {
+        Self { parent, area }
+    }
 }
 
 impl<T> OriginDimensions for SubImage<'_, T> {