+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Embedded Graphics is a `no_std` library for adding graphics features to display 
 - **(breaking)** [#571](https://github.com/embedded-graphics/embedded-graphics/pull/571) Added color argument to `MockDisplay::from_points` to make it usable for all color types.
 - **(breaking)** [#572](https://github.com/embedded-graphics/embedded-graphics/pull/572) Removed the `height` argument from `ImageRaw::new`. The `height` is now calculated based on the width and data length.
 - **(breaking)** [#572](https://github.com/embedded-graphics/embedded-graphics/pull/572) Replaced `pixelcolor::raw::RawDataIter` by the types in the `iterator::raw` module.
+- **(breaking)** [#573](https://github.com/embedded-graphics/embedded-graphics/pull/573) Changed `MonoFont` from a trait to a struct.
+- **(breaking)** [#573](https://github.com/embedded-graphics/embedded-graphics/pull/573) Renamed `ContiguousIteratorExt::into_pixels` to `into_pixels_iter` to reduce possible confusion with `IntoPixels::into_pixels`.
 
 ### Fixed